requirements.txt
eip712
eth-ape
fastapi
uvicorn
silverback
sqlmodel
taskiq-redis
